The Survey widget and feature was restructured
