+++ src/features/PassSurvey/ui/SurveyHeader.tsx
@@ -1,17 +1,19 @@
 import { useContext } from 'react';
 
-import { SurveyContext } from '~/features/PassSurvey';
+import { EntityTimer } from './EntityTimer';
+import { SurveyContext } from '../lib';
+
 import { SaveAndExitButton } from '~/features/SaveAssessmentAndExit';
 import { MultiInformantTooltip } from '~/features/TakeNow';
 import { Theme } from '~/shared/constants';
-import { AvatarBase, BaseProgressBar, Box, ClockIcon, Text } from '~/shared/ui';
-import { isStringExist, useCustomMediaQuery } from '~/shared/utils';
+import { AvatarBase, BaseProgressBar, Box, Text } from '~/shared/ui';
+import { HourMinute, isStringExist, useCustomMediaQuery } from '~/shared/utils';
 
 type Props = {
   progress?: number;
   isSaveAndExitButtonShown: boolean;
 
-  entityTimer?: string;
+  entityTimer?: HourMinute;
 };
 
 const SurveyHeader = (props: Props) => {
@@ -38,12 +40,7 @@ const SurveyHeader = (props: Props) => {
         borderBottom: `1px solid ${Theme.colors.light.surfaceVariant}`,
       }}
     >
-      {props.entityTimer && (
-        <Box display="flex" padding="8px 12px" gap="8px">
-          <ClockIcon width="24px" height="24px" color={Theme.colors.light.outline} />
-          <Text color={Theme.colors.light.outline}>{props.entityTimer}</Text>
-        </Box>
-      )}
+      {props.entityTimer && <EntityTimer entityTimerSettings={props.entityTimer} />}
       <MultiInformantTooltip />
 
       <Box sx={{ gridColumn: '2 / 3' }}>
